pyproject.toml
[tool.poetry]
name = "kra-connect"
version = "0.1.0"
description = "Official Python SDK for KRA GavaConnect API - Tax compliance and verification for Kenya"
authors = ["KRA-Connect Team <developers@kra-connect.dev>"]
license = "MIT"
readme = "README.md"
homepage = "https://kra-connect.dev"
repository = "https://github.com/your-org/kra-connect"
documentation = "https://docs.kra-connect.dev/python"
keywords = [
    "kra",
    "kenya-revenue-authority",
    "gavaconnect",
    "tax-compliance",
    "pin-verification",
    "tcc-verification",
    "kenya",
    "api-client"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
]
packages = [{include = "kra_connect", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.25.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
tenacity = "^8.2.3"
cachetools = "^5.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
black = "^23.12.0"
isort = "^5.13.0"
mypy = "^1.7.0"
pylint = "^3.0.3"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
respx = "^0.20.2"

[tool.poetry.scripts]
kra = "kra_connect.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=kra_connect --cov-report=term-missing --cov-report=html --cov-report=xml"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/kra_connect"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
